moved json into data folder
